# Conflicts:
#	rest-controller/src/main/webapp/css/default.css
#	rest-controller/src/main/webapp/html/HomeUnique.html
#	rest-controller/src/main/webapp/js/home.js

+++ controller/src/main/java/org/template/controller/UserServiceControllerBean.java
@@ -3,7 +3,6 @@ package org.template.controller;
 import org.template.interfaces.IUserServiceController;
 import org.template.model.*;
 
-import javax.ejb.Stateful;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
@@ -12,10 +11,8 @@ import javax.persistence.Query;
 import java.sql.Timestamp;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
-import java.util.Optional;
 
 @Stateless(name = "UserServiceEJB")
 public class UserServiceControllerBean implements IUserServiceController {
@@ -83,8 +80,17 @@ public class UserServiceControllerBean implements IUserServiceController {
     }
 
     @Override
-    public void deleteUserService(Integer idService) {
-        em.remove(em.getReference(Service.class, idService));
+    public void deleteUserService(Integer idService, String role) {
+        if(role.equals("R"))
+            em.remove(em.getReference(Service.class, idService));
+        else if(role.equals("P")){
+            Query q = this.em.createQuery(
+                    "update Service  set performerUser=:user where performerUser is not null and idService = :id");
+            q.setParameter("user", null);
+            q.setParameter("id",idService);
+            int quan=q.executeUpdate();
+            int aux=quan;
+        }
     }
